AER-3284, AER-6317, AER-6320 - added various record memory size features.
* AER-3284 - added object size histograms for storage-engine memory namespaces.

* AER-6317 - added record.memory_size() UDF method.

* AER-6320 - added as_exp_memory_size() operation.

+++ as/include/base/udf_record.h
@@ -67,7 +67,7 @@ typedef struct udf_record_s {
 	bool has_updates;
 
 	uint8_t result_code; // only set when we fail execute_updates()
-	uint64_t old_memory_bytes; // DIM only
+	uint32_t old_memory_bytes; // DIM only
 
 	// Non-DIM only.
 	uint8_t* particle_buf;
